PostResultado y GetPartida

+++ Ahorcado.Entidades/Partida.cs
@@ -11,7 +11,7 @@ namespace Ahorcado.Entidades
         public int Id { get; set; }
         public int gano { get; set; }
         public int intentos_disponibles { get; set; }
-        public float tiempo_transcurrido { get; set; }
+        public double tiempo_transcurrido { get; set; }
         public int cantidad_letras_adivinadas { get; set; }
     }
 }
